prefer cut to awk -n
Awk's -n, --non-decimal-data flag is intended for converting octal and
hexadecimal numeric values to decimal, but all we're doing here is
extracting the first field from the sha256sum output.

The -n wasn't actually doing anything in this case, since "print" treats
its expressions as strings by default:

  echo 'f8a496 ...' | awk -n '{print $1}'
  f8a496

  echo 'f8a496 ...' | awk '{print $1}'
  f8a496

The conversion occurs if you coerce the input field to a numeric, and
even then it will only recognize a hexadecimal with a leading "0x":

  echo '0xf8a496 ...' | awk -n '{print $1 + 0}'
  16295062

  echo '0xf8a496 ...' | awk -n '{printf "%d\n", $1}'
  16295062

  echo 'f8a496 ...' | awk -n '{print $1 + 0}'
  0

  echo 'f8a496 ...' | awk -n '{printf "%d\n", $1}'
  0

The Gawk documentation recommends aginst using this flag, as its
behavior can be surpising and break old programs badly. It recommends
using strtonum() instead. All that's irrelevant for our use-case, but I
thought it was interesting.

https://www.gnu.org/software/gawk/manual/html_node/Nondecimal-Data